Improve status bar lyric
