+++ app/src/main/java/remix/myplayer/service/notification/NotifyImpl.kt
@@ -4,7 +4,6 @@ import android.app.Notification
 import android.content.Context
 import android.graphics.Bitmap
 import android.graphics.drawable.Drawable
-import android.os.Build
 import android.view.View
 import android.widget.RemoteViews
 import androidx.core.app.NotificationCompat
@@ -165,30 +164,6 @@ class NotifyImpl(context: MusicService) : Notify(context) {
         if (service.isDesktopLyricLocked) Command.UNLOCK_DESKTOP_LYRIC else Command.TOGGLE_DESKTOP_LYRIC)
     remoteBigView.setOnClickPendingIntent(R.id.notify_lyric, lyricIntent)
     remoteView.setOnClickPendingIntent(R.id.notify_lyric, lyricIntent)
-  }
-
-  override fun updateWithLyric(lrc: String) {
-    if (!service.isPlaying) return
-    val song = service.currentSong
-    val builder = NotificationCompat.Builder(service, PLAYING_NOTIFICATION_CHANNEL_ID)
-    builder.setContentText(song.title)
-        .setContentTitle(song.title)
-        .setShowWhen(false)
-        .setTicker(lrc)
-        .setOngoing(service.isPlaying)
-        .setContentIntent(contentIntent)
-        .setSmallIcon(R.drawable.icon_notifbar)
-    val notification = builder.build()
-
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-      notification.extras.putInt("ticker_icon", R.drawable.icon_notifibar_lrc)
-      notification.extras.putBoolean("ticker_icon_switch", false)
-    }
-
-    notification.flags = notification.flags.or(FLAG_ALWAYS_SHOW_TICKER)
-    notification.flags = notification.flags.or(FLAG_ONLY_UPDATE_TICKER)
-
-    pushNotify(notification)
   }
 
 //  private fun tintBitmap(resId: Int, color: Int?): Bitmap {
