+++ app/src/main/java/remix/myplayer/service/notification/NotifyImpl24.kt
@@ -91,29 +91,4 @@ NotifyImpl24(context: MusicService) : Notify(context) {
         .build()
     pushNotify(notification)
   }
-
-  override fun updateWithLyric(lrc: String) {
-    if (!service.isPlaying) return
-    val song = service.currentSong
-    val builder = NotificationCompat.Builder(service, PLAYING_NOTIFICATION_CHANNEL_ID)
-    builder.setContentText(song.artist + " - " + song.album)
-        .setContentTitle(song.title)
-        .setShowWhen(false)
-        .setTicker(lrc)
-        .setOngoing(service.isPlaying)
-        .setContentIntent(contentIntent)
-        .setSmallIcon(R.drawable.icon_notifbar)
-    val notification = builder.build()
-
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.KITKAT) {
-      notification.extras.putInt("ticker_icon", R.drawable.icon_notifibar_lrc)
-      notification.extras.putBoolean("ticker_icon_switch", false)
-    }
-
-    notification.flags = notification.flags.or(FLAG_ALWAYS_SHOW_TICKER)
-    notification.flags = notification.flags.or(FLAG_ONLY_UPDATE_TICKER)
-
-    pushNotify(notification)
-  }
-
 }
